// MyMusic.Client/src/model/searchMetadataResult.ts
// @ts-nocheck
/**
Generated by orval 🍺
Do not edit manually.
MyMusic.Server | v1
OpenAPI spec version: 1.0.0
*/ import type { SourceSong } from "./sourceSong";

export interface SearchMetadataResult {
	/** @pattern ^-?(?:0|[1-9]\d*)$ */
	sourceId: number;
	sourceName: string;
	sourceIcon: string;
	song: SourceSong;
}